back/fastapi/glim/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
Deprecated==1.2.14
distro
docarray
docker==7.0.0
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi
filelock==3.13.1
frozenlist==1.4.1
google-api-core==2.15.0
google-auth==2.25.2
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-translate==3.13.0
google-cloud-vision==3.5.0
googleapis-common-protos==1.62.0
grpcio==1.57.0
grpcio-health-checking==1.57.0
grpcio-reflection==1.57.0
grpcio-status==1.57.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx
httpx-sse==0.4.0
idna==3.6
importlib-metadata==6.11.0
jcloud==0.3
jina
jina-hubble-sdk
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.0
langchain-community==0.0.9
langchain-core==0.1.7
langserve==0.0.38
langsmith==0.0.77
markdown-it-py==3.0.0
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.6.1
opentelemetry-api==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-exporter-otlp-proto-common==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-otlp-proto-http==1.19.0
opentelemetry-exporter-prometheus==0.41b0
opentelemetry-instrumentation==0.40b0
opentelemetry-instrumentation-aiohttp-client==0.40b0
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-grpc==0.40b0
opentelemetry-proto==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-semantic-conventions==0.40b0
opentelemetry-util-http==0.40b0
orjson==3.9.10
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
prometheus-client==0.19.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydub==0.25.1
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
colorama==0.4.6
greenlet==3.0.3
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
sse-starlette==1.8.2
starlette
tenacity==8.2.3
tqdm==4.66.1
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.9.0
uvicorn==0.23.1
urllib3
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
pydantic==2.5.3
pydantic_core==2.14.6